tx: Add savepoints support
Support for savepoints are added to `tx.Transactor` via the
`tx.WithSavepoints` option. The syntax is compatible with MariaDB,
sqlite and Postgres.

The semantics mainly affect InTx being called recursively via user code,
if a child InTx returns with an error, the side-effect (in the
transaction) of this savepoint will be reverted from the point of view
of the parent InTx.