mailbox: Fix consumer message metadata
The consumer was not scanning the metadata column of the mailbox table
when claiming a message. This commit fixes the issue by adding the
metadata column to the claim statement and scanning the value into a
`map[string]string` before converting it to a `Metadata` type. This
tested by adding a new test case to the consumer integration test.